% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\name{na}
\alias{na}
\title{: NA remove}
\usage{
na(data)
}
\arguments{
\item{data}{data frame}
}
\description{
Removes missing data
This function will remove sample rows that have any missing SNP data (NA)
}
\author{
Dayana
}
